% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{makeDashElement}
\alias{makeDashElement}
\title{Create dashed rectangle objects for legends}
\usage{
makeDashElement(
  dashArray = "none",
  id = "rect",
  strokeWidth,
  height,
  width,
  color,
  fillColor,
  opacity,
  fillOpacity,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{dashArray}{Dash pattern}
}
\value{
<rect> object string
}
\description{
Result is string of <rect> objects, a la `symbolSvg()`
}
