% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{makeSymbol}
\alias{makeSymbol}
\alias{makeSymbolIcons}
\title{Create an SVG tag for the symbol}
\usage{
makeSymbol(
  shape,
  width,
  height = width,
  color,
  fillColor = color,
  opacity = 1,
  fillOpacity = opacity,
  ...
)

makeSymbolIcons(
  shape = c("rect", "circle", "triangle", "plus", "cross", "diamond", "star",
    "stadium", "line", "polygon"),
  color,
  fillColor = color,
  opacity,
  fillOpacity = opacity,
  strokeWidth = 1,
  width,
  height = width,
  ...
)
}
\arguments{
\item{shape}{the desired shape of the symbol}

\item{width}{in pixels}

\item{height}{in pixels}

\item{color}{color of the symbol}

\item{fillColor}{fill color of symbol}

\item{opacity}{opacity of color}

\item{fillOpacity}{opacity of fillColor}

\item{...}{arguments to be passed to svg shape tag}

\item{strokeWidth}{width in pixels of symbol outline}
}
\value{
HTML svg element
}
\description{
Create an SVG tag for the symbol
}
