% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{makeSymbol}
\alias{makeSymbol}
\title{Create an SVG tag for the symbol}
\usage{
makeSymbol(shape, label, color, labelStyle, width, height)
}
\arguments{
\item{shape}{the desired shape of the symbol}

\item{label}{label to be placed to the left of the symbol}

\item{color}{fill color of the symbol}

\item{labelStyle}{character string of style argument for HTML text}

\item{width}{in pixels}

\item{height}{in pixels}
}
\value{
HTML svg element
}
\description{
Create an SVG tag for the symbol
}
