% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{mapSymbols}
\alias{mapSymbols}
\alias{makeSymbol}
\alias{makeSvgUri}
\alias{makeSymbolIcons}
\alias{makeSymbolText}
\alias{makeSymbolTextIcons}
\alias{addSymbols}
\alias{addSymbolsSize}
\alias{addText}
\alias{addTextSize}
\alias{sizeNumeric}
\alias{sizeBreaks}
\alias{makeSymbolsSize}
\title{Create Map Symbols for 'leaflet' maps}
\usage{
makeSymbol(
  shape,
  width,
  height = width,
  color,
  fillColor = color,
  opacity = 1,
  fillOpacity = opacity,
  ...
)

makeSvgUri(svg, width, height, strokeWidth)

makeSymbolIcons(
  shape,
  color,
  fillColor = color,
  opacity,
  fillOpacity = opacity,
  strokeWidth = 1,
  width,
  height = width,
  ...
)

makeSymbolText(
  text,
  width,
  height = width,
  color = "black",
  fillColor = color,
  opacity = 1,
  fillOpacity = opacity,
  fontSize = round(min(width, height) * 0.6),
  fontFamily = NULL,
  ...
)

makeSymbolTextIcons(
  text,
  color = "black",
  fillColor = color,
  opacity = 1,
  fillOpacity = opacity,
  strokeWidth = 0,
  width,
  height = width,
  fontFamily = NULL,
  ...
)

addSymbols(
  map,
  lng,
  lat,
  values,
  shape,
  color,
  fillColor = color,
  opacity = 1,
  fillOpacity = opacity,
  strokeWidth = 1,
  width = 20,
  height = width,
  dashArray = NULL,
  data = leaflet::getMapData(map),
  ...
)

addSymbolsSize(
  map,
  lng,
  lat,
  values,
  shape,
  color,
  fillColor = color,
  opacity = 1,
  fillOpacity = opacity,
  strokeWidth = 1,
  baseSize = 20,
  data = leaflet::getMapData(map),
  ...
)

addText(
  map,
  lng,
  lat,
  text,
  color = "black",
  fillColor = color,
  opacity = 1,
  fillOpacity = opacity,
  strokeWidth = 0,
  width = 20,
  height = width,
  data = leaflet::getMapData(map),
  ...
)

addTextSize(
  map,
  lng,
  lat,
  values,
  text,
  color = "black",
  fillColor = color,
  opacity = 1,
  fillOpacity = opacity,
  strokeWidth = 0,
  baseSize = 20,
  data = leaflet::getMapData(map),
  ...
)

sizeNumeric(values, baseSize)

sizeBreaks(values, breaks, baseSize, ...)

makeSymbolsSize(
  values,
  shape = "rect",
  color,
  fillColor,
  opacity = 1,
  fillOpacity = opacity,
  strokeWidth = 1,
  baseSize,
  ...
)
}
\arguments{
\item{shape}{the desired shape of the symbol, See \link[leaflegend]{availableShapes}}

\item{width}{in pixels}

\item{height}{in pixels}

\item{color}{stroke color}

\item{fillColor}{fill color}

\item{opacity}{stroke opacity}

\item{fillOpacity}{fill opacity}

\item{...}{arguments to pass to

svg shape tag for makeSymbol, makeSymbolIcons, makeSymbolsSize

svg text tag for makeSymbolText, makeSymbolTextIcons

\link[leaflet]{addMarkers} for addSymbols, addSymbolsSize, addText, addTextSize

\link[base]{pretty} for sizeBreaks}

\item{svg}{inner svg tags for symbol}

\item{strokeWidth}{stroke width in pixels}

\item{text}{character string to render inside the svg}

\item{fontSize}{size of the text in pixels; defaults to \code{round(min(width, height) * 0.6)}}

\item{fontFamily}{optional font family for the svg text; if \code{NULL} the browser default is
used}

\item{map}{a map widget object created from 'leaflet'}

\item{lng}{a numeric vector of longitudes, or a one-sided formula of the form \code{~x}
where \code{x} is a variable in \code{data}; by default
(if not explicitly provided), it will be automatically inferred from data
by looking for a column named \code{lng}, \code{long}, or \code{longitude}
(case-insensitively)}

\item{lat}{a vector of latitudes or a formula (similar to the \code{lng} argument; the
names \code{lat} and \code{latitude} are used when guessing the latitude
column from \code{data})}

\item{values}{the values used to generate shapes; can be omitted for a single type of
shape}

\item{dashArray}{a string or vector/list of strings that defines the stroke dash pattern}

\item{data}{the data object from which the argument values are derived; by default, it
is the \code{data} object provided to \code{leaflet()} initially, but can be
overridden}

\item{baseSize}{re-scaling size in pixels of the mean of the values, the average value will
be this exact size}

\item{breaks}{an integer specifying the number of breaks or a numeric vector of the breaks;
if a named vector then the names are used as labels.}
}
\value{
HTML svg element
}
\description{
Create Map Symbols for 'leaflet' maps
}
\examples{

library(leaflet)
data(quakes)
quakes1 <- quakes[1:10,]

# makeSymbolText returns a single SVG data URI
makeSymbolText(text = 'A', width = 30, color = 'red')

# makeSymbolTextIcons builds a vector of icons
iconSet <- makeSymbolTextIcons(text = LETTERS[1:nrow(quakes1)],
                               width = 30, color = 'white',
                               fillColor = 'navy')
leaflet(quakes1) \%>\%
  addTiles() \%>\%
  addMarkers(lng = ~long, lat = ~lat, icon = iconSet)

# addText draws text labels at each location
leaflet(quakes1) \%>\%
  addTiles() \%>\%
  addText(lng = ~long, lat = ~lat,
          text = ~as.character(seq_len(nrow(quakes1))),
          color = 'white', fillColor = 'red', width = 30)

# addTextSize scales text by a numeric variable
leaflet(quakes1) \%>\%
  addTiles() \%>\%
  addTextSize(lng = ~long, lat = ~lat,
              text = ~as.character(round(mag, 1)),
              values = ~mag, color = 'black', fillColor = 'yellow',
              baseSize = 30)

}
